Change log.Fatal to panic in usage example
panic() call is a better fit to handle Define() errors, as the only
errors here are the ones of program logic, not something that program
users' can fix providing different input.

The proper solution would be for Define() to panic itself and not return
anything at all, but I'm leaving it as it for the sake of backwards
compatibility as it is mostly a cosmetic issue.